单机多开可以，连接、断开连接未经仔细测试，有待添加名字设置 1.别的电脑似乎收不到消息 2.server release模式下中断，百度说八成是内存分配问题，如数组越界

+++ src/cardHeap.cpp
@@ -63,7 +63,7 @@ card cardHeap::getCard() {	//获取一张牌，这里直接获取顶部的牌
 	return this->getTopCard();
 }
 card cardHeap::getTopCard() {	//获取顶部的牌
-	if (this->topIndex >= maxCards) {
+	if (this->topIndex >= (int)(maxCards)) {
 		return card(cardColor::ErrorColor, cardNumber::ErrorNumber);
 	}
 	card res = this->cardArray[this->topIndex];
